+++ packages/nft-checkout/src/config.ts
@@ -1,7 +1,9 @@
 import { Config } from './types'
 
 export const DEFAULT_CONFIG: Config = {
-  V3_TOKEN_LIST: 'https://ipfs.io/ipns/tokens.uniswap.org',
-  V2_TOKEN_LIST:
-    'https://raw.githubusercontent.com/traderjoe-xyz/joe-tokenlists/main/src/joe.tokenlist-v2.json',
+  UNISWAP_V3_TOKEN_LIST_URL: 'https://ipfs.io/ipns/tokens.uniswap.org',
+  TRADER_JOE_TOKEN_LIST_URL:
+    'https://raw.githubusercontent.com/traderjoe-xyz/joe-tokenlists/main/mc.tokenlist.json',
+  PANCAKE_SWAP_TOKEN_LIST_URL:
+    'https://tokens.pancakeswap.finance/pancakeswap-extended.json',
 }