+++ packages/nft-checkout/src/errors/index.ts
@@ -1,0 +1,6 @@
+import * as operatorErrors from './operator'
+export * from './operator'
+
+export const errors = {
+  ...operatorErrors,
+}