+++ packages/nft-checkout/src/types/index.ts
@@ -1,2 +1,4 @@
 export * from './operation'
 export * from './common'
+export * from './token'
+export * from './config'