+++ packages/provider/src/errors/index.ts
@@ -1,7 +1,8 @@
 import * as providerErrors from './provider'
-import * as operatorErrors from './operator'
 
 export const errors = {
   ...providerErrors,
-  ...operatorErrors,
 }
+
+export * from './runtime'
+export * from './provider'